Finished assignment, daniele tomerini
Added the files required to create the postgres database tables, insertion, and
plotting of the results.

I added some readme files in the directories I modified to explain the rationale
between the decision

Pip file was changed to include the string distance calculator and a package to
fix excel file reading. I did not fix the version